adds in api and  views directory and index.ts for exports
