+++ src/views/MainView/MainView.modules.css
@@ -1,0 +1,3 @@
+.mainView{
+    padding: 20px;
+}
